Show version and help #31
