+++ src/controller/mod.rs
@@ -1,0 +1,1 @@
+pub(super) mod controller_main;
