--- src/handler.rs
@@ -1,14 +1,0 @@
-use crate::{fuzzy_finder, parser::makefile::Makefile};
-use std::process;
-
-pub fn run() {
-    let makefile = match Makefile::create_makefile() {
-        Err(e) => {
-            println!("[ERR] {}", e.to_string());
-            process::exit(1)
-        }
-        Ok(f) => f,
-    };
-
-    fuzzy_finder::run(makefile);
-}
