--- src/parser/mod.rs
@@ -1,3 +1,0 @@
-pub(super) mod include;
-pub(super) mod makefile;
-pub(super) mod target;
