+++ src/usecase/usecase.rs
@@ -1,0 +1,5 @@
+pub trait Usecase {
+    fn command_str(&self) -> Vec<&'static str>;
+    // TODO: Return Result.
+    fn run(&self);
+}
